Add explicit and make public headers compile with -Wzero-as-null-pointer-constant
... or equivalent.

QtBase 5.6 headers already compile that way, so let the other
modules follow suit.

Cleaned up placement of * and & in parameters as a drive-by.

Added explicit where it was missing as a drive-by. This is
not a source-incompatible change, because code that breaks
by this is a bug. Let's not have this sitting around in an LTS.

Task-number: QTBUG-45291
Change-Id: If81ed0c71393aee21d347f5ade4bf3fcc07cd82f
Reviewed-by: Yoann Lopes <yoann.lopes@theqtcompany.com>