Remove AndroidManifest.xml from qmlvideofx example.
The manifest in qmlvideofx example is useless, and with the strict
requirement of API level 16 or higher (585287d0@qttools), it means it's
not working out of the box anymore. To avoid similar issues in the
future, this change just removes the manifest (instead of bumping the
min. version).

Task-number: QTBUG-52374
Change-Id: I6c672bdbe23935852c87c93e425003b7b6827771
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>