Add config test for wasapi plugin
The wasapi audio plugin can also be used on desktop versions of Windows.
The support has been experimental so far and will now be enabled in case
the SDK contains support for it.

Unfortunately MinGW does not provide Microsoft::WRL::RuntimeClass and
IActivateAudioInterfaceCompletionHandler, hence we need to disable the
plugin currently for it.

Change-Id: I2be6e99810d462507112465a2084a2d72c3c62f0
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>