Fix playing a playlist containing an invalid media as first item
When a playlist contains an invalid media, QMediaPlayer normally
continues to the next media. If an invalid media was first in the list,
the backend would never go into PlayingState, which would then cause
the media player to never start the next item in the list (thinking
it's stopped).
We now always transition to PlayingState when starting playback of a
playlist, regardless of the backend going into that state or not.

Change-Id: I4227f937e7a619afbd8adbe2fccedfa5d43ad89a
Reviewed-by: Christian Stromme <christian.stromme@qt.io>