winrt: Make focus actions synchronous
The camera and its controls have been moved out of the xaml thread
recently, hence acting asynchronously is not required anymore for the
focus control. In addition to this, previous behavior caused troubles
inside QML as QDeclarativeCameraFocus::setFocusMode always emitted a
focus change even though it did not happen yet. This resulted in
inconsistencies between QML and the plugin.

Task-number: QTBUG-48539
Change-Id: I19a3a3512530b01b627476e233291d737cfee11f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Samuel Nevala <samuel.nevala@intopalo.com>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>