Stabilize tst_QAudioOutput::pullSuspendResume()
Don't wait after resume(); the test is meant to check the state of the
QAudioOutput right after the call, not later. The state and the error
status are supposed to change immediately, even if the backend runs in
another thread.
Ignore any state change that can happen because of underrun errors
(likely to happen on CI).
Removed the test from blacklist.

Change-Id: Idaef6a9e0dfcfc89424fb2383cf3830a3184f975
Reviewed-by: Milla Pohjanheimo <milla.pohjanheimo@qt.io>
Reviewed-by: Christian Stromme <christian.stromme@qt.io>