WindowsAudio: fix volume when using multiple QAudioOutputs
We were using Windows' waveOutSetVolume() function to set the volume
on an audio stream. That function applies the volume on the sound
device and not on a particular stream, meaning all QAudioOutputs and
QSoundEffects were sharing the same volume. All Qt audio streams are
supposed to have their volume controllable separately; in order to
respect that, we now apply the volume ourselves on the data before
writing it to the device.

Task-number: QTBUG-43765
Task-number: QTBUG-48049
Change-Id: I407588ca40cf4216454431fe3e6ce4d8740f391d
Reviewed-by: Christian Stromme <christian.stromme@qt.io>