winrt: Fix pause/play for videos
This is partially a revert of bff19dbe6733d245. While that commit fixed
switching cameras it did not respect pausing a video, where setActive is
called from a different thread than in the camera case. Using shutdown
introduced a wait, which in that case deadlocked.

Instead revert to request an interrupt, but when restarting make sure
that the render thread is on hold before start is invoked.

Task-number: QTBUG-53722
Change-Id: Id5a722911bb2c30509fabb73bec925cadbf77628
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>