Disable the ALSA plugin for QNX
The QNX ALSA implementation isn't compatible with the generic ALSA
plugin.  QNX has its own ALSA plugin.

The old configure system didn't test for ALSA support on QNX, but
the new configure system does test for it.  A porting error.  With
an official QNX SDP, this isn't a problem, because the ALSA headers
aren't in the expected location, so the test fails.  With some
QNX internal SDPs though, the ALSA headers are in the expected
location, so the test succeeds.

Change-Id: Id62aa4c17c683397545bf42fc5b741f25b719360
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>