Limit the size of the QnxAudioOutput stack buffer
QnxAudioOutput allocates a buffer on the stack based on the free value
from snd_pcm_plugin_status, but the way that QnxAudioOutput configures
the stream, how QnxAudioOutput currently pauses playback, and a bug in
io-audio combine to cause io-audio to produce very large free values
when resuming playback after a long pause.  As a result, QnxAudioOutput
allocates a stack buffer that causes a stack overflow.  Allocating
a buffer on the stack with a size that isn't restrained in any way
isn't a good idea.  Put some constraints on the size.

Change-Id: I2b72e72504041f0caeb591912662fb9bed931b21
Reviewed-by: Dan Cape <dcape@qnx.com>
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>