Implement QAudioRoleControl for QNX
The system must provide a JSON file that maps Qt audio roles to
the system's audio types.  This is necessary because QNX doesn't
dictate the audio types that a system must use.  Anyone creating
a QNX system is free to define whatever audio types they deem to
be appropriate for that system.

Set the audio role for the Qt player example to VideoRole and
added code to dump the supported audio roles.

Change-Id: I34d2c5da0033f69b9dd476a0eadccf3d87d07bfd
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>