Switch the metadata code to the event API
The switch is necessary because Multimedia for QNX 7.0.0 removes
support for PPS based retrieval.  PPS itself is deprecated in QNX
7.0.0.  QNX 6.6.0 is also switched to using event notification
since Multimedia for QNX 6.6.0 also supports it.

There's a slight change in when metadata first becomes available
because an event is used instead of file based retrieval.  I think
it's a necessary change because there isn't actually any guarantee
that the metadata is complete at the time that the code used to
request it and the code may have missed changes because it never
read the information again.

[ChangeLog][QNX] Switch to mmr_event_t based metadata retrieval.
PPS based retrieval is not supported by Multimedia for QNX 7.0.0.

Change-Id: I2b70f05422ee03d25ed2446a0e30b56b03dd82c8
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Christian Stromme <christian.stromme@qt.io>