Prevent m_runningCount from eventually overflowing
When m_loopCount is Infinite (i.e. -1) m_runningCount would go from -2
and down, eventuall wrapping and reaching 0 and thus stopping.

This is a theorical problem since even if we had a 1 second video it would
take lots of years to trigger but i guess it's better than
having the value of m_runningCount be unbounded

Change-Id: I340ec2157ece3334e58ab4afd77c34b478a256e9
Reviewed-by: Christian Stromme <christian.stromme@qt.io>