PulseAudio: allow to specify the pulseaudio server string
Currently we call pa_context_connect() with an empty server string.
This works fine if we only want to connect to the default pulseaudio
server, but we may want to specify the server string on systems with
custom configuration.

Add a new environment variable: QT_PULSE_SERVER_STRING which allows to
specify the server string passed to pa_context_connect().

Change-Id: I6805ff5e941d13bc91d306e5df3ab5d04a8a90af
Reviewed-by: Christian Stromme <christian.stromme@qt.io>