winrt: Use QMutex instead of criticalSections
critical sections are really unpredictable when it
comes to order of activation. It was possible that
the loop in QWinRTAbstractVideoRendererControl's
syncAndRender blocked its setBlitMode for up to 30
seconds on application start. During this time the
camera screen just stayed black. The whole approach
seems to work a lot better when QMutexes are used.

Task-number: QTBUG-63015
Change-Id: Ib1b0fa1da35fe299896068146254e4cf1d9616fb
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>