DirectShow: Emit an error if QCamera::start() fails
If a camera is not started successfully,
then an error() should be emitted.

After an error the camera's state will be QCamera::UnloadedState
and status will be QCamera::UnloadedStatus.

The error signal is handled when the camera is unable to set following states:
QCamera::UnloadedState, QCamera::LoadedState or QCamera::LoadingState.

Thus additionally to QCamera::start() an error can be emitted
even when QCamera::load(), QCamera::unload(), or QCamera::stop() is
called.

Task-number: QTBUG-51825
Change-Id: Ib5ea08ed7983ea49a7bf8c0321cc5266a68d9144
Reviewed-by: Christian Stromme <christian.stromme@qt.io>