DirectShow: Use pixel aspect ratio
Originally pixel aspect ratio was clearing (set to 1,1)
within renegotiating the media type.
Which caused samples with not standard pixel aspect ratio (1,1)
to be rendered with black squares on the top and the bottom.
(Original size was kept but PAR cleared)

Added a fix to keep PAR while adjusting the media type to optimal one
and set original PAR to QVideoSurfaceFormat.

Task-number: QTBUG-64438
Change-Id: I6d1a5c441b89bbd5cea86667cc570bb7eb19068b
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>