AVFoundation: Use 90 for the front facing camera's orientation
Since the information is not available as to what the orientation is
from the AVFoundation API then we need to rely on common sense. Since
the back facing camera is 270, then it stands to reason that the front
one would be 90. This has been tested on macOS and iOS and all three
cameras behave correctly.

Task-number: QTBUG-37955
Change-Id: I1ab079cb5d4337948541e58321df51efbadf3e39
Reviewed-by: Andy Nichols <andy.nichols@qt.io>