PulseAudio: Fix deadlock for QSoundEffect in setMuted, setVolume
The deadlock would happen when one of the protected functions were
reentered, as the lock used was not recursive.

Since reading and writing to the volume and muted properties don't
happen that often, the rw-mutex was replaced by a normal non-recursive
mutex, which is now unlocked before the property changed signal is emitted.

Task-number: QTBUG-65220
Change-Id: I1a224e9ff03f14593a854a7c1049c8a0445e29e6
Reviewed-by: Christian Stromme <christian.stromme@qt.io>