WinRT: Read data in callback to prevent deadlock
When playing mp3 with QMediaPlayer and seeking is performed (IMFMediaEngineEx::SetCurrentTime)
and also IMFByteStream::BeginRead is already done (actual data has not
been read yet) but IMFByteStream::EndRead is not called.
There would be a possibility to get a deadlock when the playback gets stuck.

Possible problem call flow:
1. BeginRead is called
2. Queue finishRead and postpone it
3. SetCurrentTime to new position (is called on event loop)
4. finishRead is called
5. EndRead is called

As a result MF_MEDIA_ENGINE_EVENT_SEEKING is received but
MF_MEDIA_ENGINE_EVENT_SEEKED is never received.

Proposing to read data directly from BeginRead callback and avoid
possibility to seek when data is not read yet.

Task-number: QTBUG-67614
Change-Id: I809e5df1c05de8014ebd0a48fd008a291f39433c
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>